working on a news cell

+++ Stockify/Models/NewsStory.swift
@@ -9,7 +9,7 @@ import Foundation
 
 struct NewsStory: Codable {
     let category: String
-    let datetime: Int
+    let datetime: Double
     let headline: String
     let id: Int
     let image: String
